Synced shader name between the content and options tab
And moved shaderlist and shader cycling to options as well
