+++ src/content/index.tsx
@@ -17,62 +17,12 @@ const App: React.FC = () => {
     // Local states
     const [showShaderName, _] = useState<boolean>(true);
     const [analyser, setAnalyser] = useState<AnalyserNode | undefined>();
-    const [shaderName, setShaderName] = useState<string>('MusicalHeart.frag');
-    const [shaderList, setShaderList] = useState<string[]>([]);
-    const [shaderIndex, setShaderIndex] = useState<number>(0);
     const analyserCanvasRef = useRef<HTMLCanvasElement>(null);
     const renderCanvasRef = useRef<HTMLCanvasElement>(null);
     //const orthoCamRef = useRef<OrthographicCamera>();
 
     // Synced states
-    const [reduxShaderName, setReduxShaderName] = useSyncSetState('shadername', 'MusicalHeart.frag');
-
-    const cycleShaders = () => {
-        if (shaderList.length == 0) {
-            return;
-        }
-        const newShaderName = shaderList[shaderIndex];
-        setShaderName(newShaderName);
-        const newShaderIndex = (shaderIndex + 1) % shaderList.length;
-        setShaderIndex(newShaderIndex);
-     
-        // our sync test
-        setReduxShaderName(newShaderName);
-    }
-
-
-    const fetchShaderList = async () => {
-        const shaders = await loadShaderList();
-        setShaderList(shaders);
-    }
-
-    // Initial shader list retrieval
-    useEffect(() => {
-        fetchShaderList().catch(console.error);
-    }, []);
-
-    const handleKeyUp = (e:any) => {
-        const evt = e as KeyboardEvent;
-        if (e.code != 'Space') {
-            return;
-        }
-        cycleShaders();
-    };
-
-    // Shader cycle input logic
-    // Work-around/hack to get the update state in the listener
-    useEffect(() => {
-        browser.runtime.onMessage.addListener(async (msg, sender) => {
-            if (msg.command && (msg.command === SPACE)) {
-                cycleShaders();
-            }
-        });
-        document.removeEventListener('keyup', handleKeyUp);
-        document.addEventListener('keyup', handleKeyUp);
-        return () => {
-            document.removeEventListener('keyup', handleKeyUp);
-        };
-    }, [shaderList, shaderName, shaderIndex]);
+    const [shaderName] = useSyncSetState('shadername', 'MusicalHeart.frag');
 
     const initializeAnalyzer = async () => {
         const currentTab = await getCurrentTab();
@@ -126,9 +76,6 @@ const App: React.FC = () => {
                    loop autoPlay style={{visibility: analyser ? 'hidden' : 'visible'}}></video>
             <div className="fixed flex w-screen h-screen z-[100] bg-white-200">
                 {showShaderName && <h1 className="m-2 text-2xl font-medium leading-tight text-white fixed z-40">{shaderName}</h1>}
-            </div>
-            <div className="fixed flex w-screen h-screen z-[100] bg-white-200 top-[25px]">
-                {shaderName && <h1 className="m-2 text-2xl font-medium leading-tight text-teal-100 fixed z-40">{reduxShaderName}</h1>}
             </div>
         </div>
     );
